Leftover of bigint

+++ sql-files/upgrades/upgrade_svn17600.sql
@@ -1,1 +1,1 @@
-ALTER TABLE `login` ADD `bank_vault` BIGINT( 64 ) NOT NULL DEFAULT '0';
+ALTER TABLE `login` ADD `bank_vault` int(11) NOT NULL DEFAULT '0';
